requirements.txt
flask >= 1.1
Flask-SQLAlchemy
flask-login
eventlet == 0.30.0
Flask-SocketIO==5.0.1